refactor: extracts balance sorting composables
